requirements.txt
aiohttp==3.10.11
aiosignal==1.3.1
asgiref>=3.7.2
Django>=5.1.3
dj-database-url>=2.3.0
whitenoise[brotli]>=6.0.0
async-timeout==4.0.2
attrs==23.1.0
certifi>=2024.7.4
charset-normalizer>=3.1.0
defusedxml==0.7.1
diff-match-patch==20230430
django-crispy-forms==1.14.0
django-filter==23.2
django-import-export==3.2.0
django-redis==5.2.0
django-storages==1.13.2
et-xmlfile==1.1.0
frozenlist==1.3.3
gunicorn==22.0.0
idna>=2.10
MarkupPy==1.14
multidict==6.0.4
odfpy==1.4.1
openai>=0.27.8
openpyxl==3.1.2
pillow>=10.2.0
psycopg2-binary>=2.9.9
PyMySQL==1.1.1
pytz==2023.3
PyYAML==6.0.1
redis==4.5.5
requests>=2.25.1
sqlparse==0.5.0
supabase-py>=0.0.2
tablib==3.4.0
tqdm==4.66.3
typing_extensions>=4.6.3
urllib3>=1.26.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.17.2
markdown
djangorestframework-simplejwt==5.3.1
# django-heroku
django-webpack-loader
django-cors-headers==4.3.1
python-dotenv>=1.0.0
